VPC service project connections could only be created on a second terraform run.
Using local values, and not adding depends_on lifecycle hooks breaks
terraform when it tries to create VPC service project connections,
because the list of projects is calucated before they are created.

Also, calculate the number of service project connections to create
rather than forcing the user to do it themselves.