I have had a very long, hard week with Google Cloud IAM.
Follow this document to the letter:
https://cloud.google.com/kubernetes-engine/docs/how-to/cluster-shared-vpc#console_3

Understand this document, too:
https://cloud.google.com/vpc/docs/provisioning-shared-vpc#gcloud

This document is also helpful:
https://cloud.google.com/kubernetes-engine/docs/troubleshooting#gke_service_account_deleted

Read documentation like this very carefully and realize that ANY use of
"*_iam_binding or *_iam_policy" is DESTRUCTIVE:
https://registry.terraform.io/providers/hashicorp/google/latest/docs/resources/compute_subnetwork_iam

Finally, realize that there are default service accounts that Google
lays out when you enable APIs.  Use them.